Fix FAQ items expanding behaviour
This is a fix for issue #195.

The simple idea is that the height of the faq-items automatically scaled to fill the extra height made by the neighboring faq-item.

So now by default their height is hard-coded to be the line height plus padding (1 rem above and below).

For non-active items, set max-height to this constant height, and for active items unset this limit.

Did simple tests, and the UI didn’t seem to break in any way.